% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{center_rm}
\alias{center_rm}
\title{Center expression intensities}
\usage{
center_rm(rm)
}
\arguments{
\item{rm}{data.table of smoothed expression intensities counts}
}
\value{
centered_rm data.table of smoothed, centered expression intensities
}
\description{
Take in a data.table of genomic positions and smoothed expression intensities counts and center by
subtracting average intensity across all beads for each gene
}
