% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clone_so}
\alias{clone_so}
\title{Add clone information to meta data of seurat object}
\usage{
clone_so(so, hcl_sub, md, mal = FALSE)
}
\arguments{
\item{hcl_sub}{hierarchical clustering object of cluster assignemnt as outputted from SlideCNA::plot_clones()}

\item{md}{data.table of metadata of each bead}

\item{mal}{TRUE if only using malignant beads}

\item{data}{so seurat object of beads and their meta data}
}
\value{
A seurat object updated with clone information
}
\description{
This function adds another column for cluster designation to a seurat object's meta data
}
