% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{cnv_heatmap}
\alias{cnv_heatmap}
\title{Plot CNV scores on a heat map}
\usage{
cnv_heatmap(
  cnv_data,
  md,
  chrom_colors,
  hc_function = "ward.D2",
  plot_directory
)
}
\arguments{
\item{cnv_data}{list object of cnv data from SlideCNA::prep_cnv_dat()}

\item{md}{data.table of metadata of each bead}

\item{chrom_colors}{vector of colors labeled by which chromosome they correspond to}

\item{hc_function}{character for which hierarchical clustering function to use}

\item{plot_directory}{output plot directory path}
}
\description{
This function prepares data for plotting and makes a heat map of CNV scores per bead across all genes
}
