% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{dat_to_long}
\alias{dat_to_long}
\title{Convert data to long format and add in metadata}
\usage{
dat_to_long(dat, md)
}
\arguments{
\item{dat}{data.table of smoothed relative expression intensities}

\item{md}{data.table of metadata per bead}
}
\value{
A data.table of bead expression intensities per gene with metadata in long format
}
\description{
This function will create rows for each bead and gene combination, adding in new metadata with bin designations
}
