% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{find_cluster_markers}
\alias{find_cluster_markers}
\title{Find and plot top n DEGs per cluster}
\usage{
find_cluster_markers(
  so_clone,
  type,
  logfc.threshold = 0.2,
  min.pct = 0,
  only.pos = TRUE,
  n_markers,
  text_size,
  title_size,
  legend_size_pt,
  p_val_thresh = 0.05,
  bin = FALSE,
  plotDir = None
)
}
\arguments{
\item{type}{character string that is 'all' if using malignant and normal clusters and 'malig'
if just using malignant clusters}

\item{logfc.threshold}{numeric float that is seurat parameter,
representing the minimum log2 fold change for DEGs to be significant}

\item{min.pct}{numeric Seurat function parameter}

\item{only.pos}{TRUE if only using DEGs with positive log2 fold change}

\item{n_markers}{integer of number of top DEGs to plot/use}

\item{text_size}{integer of text size for ggplot}

\item{title_size}{integer of title size for ggplot}

\item{legend_size_pt}{integer of legend size for ggplot}

\item{p_val_thresh}{threshold for p_val_adj}

\item{bin}{TRUE if using binned beads seurat object}

\item{plotDir}{output plot directory path}

\item{so}{seurat object with clone information}
}
\value{
A list object with cluster marker information
markers_clone = data.table of all cluster markers
top_markers_clone = data.table of just top cluster markers
top_clone_vis = data.frame formatted for plot visualization of top cluster markers
}
\description{
This function uses Seurat's marker finding capability to find DEGs of each cluster
}
