% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{find_go_terms}
\alias{find_go_terms}
\title{Find and plot top n GO-enriched terms per cluster}
\usage{
find_go_terms(
  cluster_markers_obj,
  type,
  n_terms = 5,
  text_size,
  title_size,
  plotDir
)
}
\arguments{
\item{cluster_markers_obj}{list object with cluster marker information}

\item{type}{character string that is 'all' if using malignant and normal clusters and 'malig'
if just using malignant clusters}

\item{n_terms}{integer of number of top DEGs to plot/use}

\item{text_size}{integer of text size for ggplot}

\item{title_size}{integer of title size for ggplot}

\item{plotDir}{output plot directory path}
}
\value{
A list object with cluster GO term information
en_clone = data.table of cluster GO terms
top_en_clone = data.table of just top cluster GO terms
}
\description{
This function utilizes cluster-specific DEGs to identify cluster-specifc GO biological processes
and plots these if they occur
}
