% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{long_to_bin}
\alias{long_to_bin}
\title{Convert to wide bin x genes + metadata format}
\usage{
long_to_bin(dat_long, plot_directory, spatial = TRUE)
}
\arguments{
\item{dat_long}{data.table of bead expression intensities per gene with metadata in long format}

\item{plot_directory}{output plot directory path}

\item{spatial}{True if using spatial information}
}
\value{
data.table of expression intensities at aggregated bin level
}
\description{
This function will combine beads into bins, taking the average expression intensities, average positions,
most common cluster seurat cluster, and most common cluster/tissue type of constituent beads
}
