% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{make_seurat_annot}
\alias{make_seurat_annot}
\title{Make a Seurat object}
\usage{
make_seurat_annot(
  cb,
  md,
  seed_FindClusters = 0,
  seed_RunTSNE = 1,
  seed_RunUMAP = 42
)
}
\arguments{
\item{cb}{counts matrix}

\item{md}{data.frame of metadata for Seurat object}

\item{seed_FindClusters}{seed for Seurat's FindClusters() function}

\item{seed_RunTSNE}{seed for Seurat's RunTSNE() function}

\item{seed_RunUMAP}{seed for Seurat's RunUMAP() function}
}
\value{
A Seurat object created from the counts matrix with the corresponding metadata
}
\description{
Using a counts matrix and meta data, create a Seurat object with defined annotations
}
