% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{make_so_bin}
\alias{make_so_bin}
\title{Make a binned version of a Seurat object}
\usage{
make_so_bin(so, md, hcl_sub, mal = FALSE)
}
\arguments{
\item{so}{Seurat object of beads and their meta data}

\item{md}{data.frame of metadata for Seurat object}

\item{hcl_sub}{hierarchical clustering object of cluster assignemnt as outputted from SlideCNA::plot_clones()}

\item{mal}{TRUE if using malignant beads only}
}
\value{
A Seurat object with binned beads as units and corresponding binned metadata
}
\description{
Aggregate Seurat object counts by bin to create a new Seurat object with binned beads
as units instead of beads
}
