% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{mean_cnv_plot}
\alias{mean_cnv_plot}
\title{Plot mean CNV scores per bin and per chromosome}
\usage{
mean_cnv_plot(
  cnv_data,
  text_size,
  title_size,
  legend_height_bar,
  plot_directory
)
}
\arguments{
\item{cnv_data}{list object of cnv data from SlideCNA::prep_cnv_dat()}

\item{text_size}{integer of text size for ggplot}

\item{title_size}{integer of title size for ggplot}

\item{legend_height_bar}{integer of bar height of legend for ggplot}

\item{plot_directory}{output plot directory path}
}
\value{
No return value, called for side effects
}
\description{
This function colors and plots each bin by its mean CNV score
on spatial coordinates for each chromosome
}
