% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{mode}
\alias{mode}
\title{Subfunction of long_to_bin() that finds mode of vector/column}
\usage{
mode(x)
}
\arguments{
\item{x}{vector (column in data.table) to calculate the mode from}
}
\value{
mode of the vector
}
\description{
This function finds the mode of a vector
}
