% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{plot_clones}
\alias{plot_clones}
\title{Plot cluster/clone information}
\usage{
plot_clones(
  cnv_data,
  md,
  k,
  type,
  chrom_colors,
  text_size,
  title_size,
  legend_size_pt,
  legend_height_bar,
  hc_function = "ward.D2",
  plotDir,
  spatial = TRUE
)
}
\arguments{
\item{md}{data.table of metadata of each bead}

\item{k}{integwer of number of clusters/clones}

\item{type}{character string, being "all" if using all binned beads, or "malig" if just malignant binned beads}

\item{hc_function}{character string for which hierarchical clustering function to use}

\item{plotDir}{output plot directory path}

\item{spatial}{TRUE if using spatial information}

\item{data}{cnv_data list object of cnv data from package::prep_cnv_dat()}
}
\value{
A hierarchical clustering object of the clusters
}
\description{
This function plots cluster dendrograms, spatial assignment, and the CNV heat map
}
