% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep}
\alias{prep}
\title{Infercnv-based preparation of relative gene expression intensities}
\usage{
prep(so, normal_beads, gene_pos, chrom_ord, logTPM = FALSE)
}
\arguments{
\item{so}{Seurat object of Slide-seq data with raw counts}

\item{normal_beads}{vector of names of normal beads}

\item{gene_pos}{data.table with columns for GENE, chr, start, end, rel_gene_pos (1 : # of genes on chromosome)}

\item{chrom_ord}{vector of the names of chromosomes in order}

\item{logTPM}{TRUE if performing adjustment with logTPM}
}
\value{
A data.table of normalized, capped, and ref-adjusted counts with genomic psoition info
}
\description{
This function takes in a data table of raw counts and a vector of reference/normal beads
to normalize counts and adjust for reference expression.
}
