% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{:=}
\title{Pipe}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{:=}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

