% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{run_enrichr}
\alias{run_enrichr}
\title{Subfunction to get significantly enriched GO terms given a set of signfiicant beads and genes}
\usage{
run_enrichr(genes, n_genes)
}
\arguments{
\item{genes}{vector of differentially expressed genes}

\item{n_genes}{number of the most significantly enriched DEGs to base gene enrichment from}
}
\value{
A data.table of the most significant GO terms and their meta data
}
\description{
This function finds the GO biological processes associated with the top n genes using enrichR
}
