% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{scalefit}
\alias{scalefit}
\title{Subfunction for scale_nUMI that normalizes a given bin for UMI count and centers the  mean CNV score at 1}
\usage{
scalefit(obj, nbin, start, end)
}
\arguments{
\item{obj}{data.table of relative expression intensities per bin}

\item{nbin}{nUMIs in that specific bin}

\item{start}{lower bound of CNV scores}

\item{end}{upper bound of CNV scores}
}
\value{
vector of adjusted CNV scores for that bins with nbin number of nUMIs within the range (inclusive) of start to end
}
\description{
This function re-scales expression intensities to be in a smaller range, normalizes for nUMI per bin,
and centers the CNV scores to have a mean of 1
}
