% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{SpatialPlot}
\alias{SpatialPlot}
\title{Spatial plots of meta data}
\usage{
SpatialPlot(
  dat_long,
  vars = NULL,
  text_size,
  title_size,
  legend_size_pt,
  legend_height_bar,
  plot_directory
)
}
\arguments{
\item{dat_long}{data.table of bead expression intensities per gene with metadata in long format}

\item{vars}{character vector of features to plot/columns of metadata}

\item{text_size}{Ggplot2 text size}

\item{title_size}{Ggplot2 title size}

\item{legend_size_pt}{Ggplot2 legend_size_pt}

\item{legend_height_bar}{Ggplot2 legend_height_bar}

\item{plot_directory}{output plot directory path}
}
\description{
This function will plot information about beads and bins on x and y coordinates
}
