% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{make_seurat_annot}
\alias{make_seurat_annot}
\title{Creation of Seurat object}
\usage{
make_seurat_annot(
  cb,
  md = NULL,
  seed_FindClusters = 0,
  seed_RunTSNE = 1,
  seed_RunUMAP = 42
)
}
\arguments{
\item{cb}{sparse counts matrix (genes x cells/beads)}

\item{md}{data.frame of meta data for cells/beads if specific annotations known}

\item{seed_FindClusters}{seed number for FindCLusters}

\item{seed_RunTSNE}{seed number for RunTSNE}

\item{seed_RunUMAP}{seed number for RunUMAP}
}
\value{
A Seurat object with specific Seurat features run
}
\description{
This function takes in raw counts (and potentially meta data) to make a Seurat object
and process it
}
