% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{mean_cnv_plot}
\alias{mean_cnv_plot}
\title{Plot mean CNV scores per bin and per chromosome}
\usage{
mean_cnv_plot(cnv_data, text_size, title_size, legend_height_bar, plotDir)
}
\arguments{
\item{cnv_data}{list object of cnv data from SlideCNA::prep_cnv_dat()}

\item{plotDir}{output plot directory path}
}
\description{
This function colors and plots each bin by its mean CNV score
on spatial coordinates for each chromosome
}
