% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{quantile_plot}
\alias{quantile_plot}
\title{Plot CNV score quantiles per bin and per chromosome}
\usage{
quantile_plot(
  cnv_data,
  cluster_label = "seurat_clusters",
  text_size,
  title_size,
  legend_height_bar,
  plotDir
)
}
\arguments{
\item{cnv_data}{list object of cnv data from package::prep_cnv_dat()}

\item{cluster_label}{character string of which column name to keep}

\item{plotDir}{output plot directory path}
}
\description{
This function colors and plots each bin by its CNV score quantiles (min, 1st quartile, median, 3rd quartile, max)
on spatial coordinates for each chromosome
}
